+++ pdfcrowd.php
@@ -387,7 +387,7 @@ Possible reasons:
 
     private $fields, $scheme, $port, $api_prefix, $curlopt_timeout;
 
-    public static $client_version = "5.6.0";
+    public static $client_version = "5.6.1";
     public static $http_port = 80;
     public static $https_port = 443;
     public static $api_host = 'pdfcrowd.com';
@@ -547,7 +547,7 @@ You need to restart your web server after installation.';
         $this->reset_response_data();
         $this->setProxy(null, null, null, null);
         $this->setUseHttp(false);
-        $this->setUserAgent('pdfcrowd_php_client/5.6.0 (https://pdfcrowd.com)');
+        $this->setUserAgent('pdfcrowd_php_client/5.6.1 (https://pdfcrowd.com)');
 
         $this->retry_count = 1;
         $this->converter_version = '20.10';
@@ -592,7 +592,7 @@ You need to restart your web server after installation.';
 
     private static $SSL_ERRORS = array(35, 51, 53, 54, 58, 59, 60, 64, 66, 77, 80, 82, 83, 90, 91);
 
-    const CLIENT_VERSION = '5.6.0';
+    const CLIENT_VERSION = '5.6.1';
     public static $MULTIPART_BOUNDARY = '----------ThIs_Is_tHe_bOUnDary_$';
 
     private function add_file_field($name, $file_name, $data, &$body) {
@@ -1178,12 +1178,12 @@ class HtmlToPdfClient {
     /**
     * Set the output page width. The safe maximum is <span class='field-value'>200in</span> otherwise some PDF viewers may be unable to open the PDF.
     *
-    * @param width Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param width The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setPageWidth($width) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $width))
-            throw new Error(create_invalid_value_message($width, "setPageWidth", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_page_width"), 470);
+            throw new Error(create_invalid_value_message($width, "setPageWidth", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_page_width"), 470);
         
         $this->fields['page_width'] = $width;
         return $this;
@@ -1192,12 +1192,12 @@ class HtmlToPdfClient {
     /**
     * Set the output page height. Use <span class='field-value'>-1</span> for a single page PDF. The safe maximum is <span class='field-value'>200in</span> otherwise some PDF viewers may be unable to open the PDF.
     *
-    * @param height Can be -1 or specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param height The value must be -1 or specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setPageHeight($height) {
         if (!preg_match("/(?i)^0$|^\-1$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $height))
-            throw new Error(create_invalid_value_message($height, "setPageHeight", "html-to-pdf", "Can be -1 or specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_page_height"), 470);
+            throw new Error(create_invalid_value_message($height, "setPageHeight", "html-to-pdf", "The value must be -1 or specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_page_height"), 470);
         
         $this->fields['page_height'] = $height;
         return $this;
@@ -1206,8 +1206,8 @@ class HtmlToPdfClient {
     /**
     * Set the output page dimensions.
     *
-    * @param width Set the output page width. The safe maximum is <span class='field-value'>200in</span> otherwise some PDF viewers may be unable to open the PDF. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
-    * @param height Set the output page height. Use <span class='field-value'>-1</span> for a single page PDF. The safe maximum is <span class='field-value'>200in</span> otherwise some PDF viewers may be unable to open the PDF. Can be -1 or specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param width Set the output page width. The safe maximum is <span class='field-value'>200in</span> otherwise some PDF viewers may be unable to open the PDF. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
+    * @param height Set the output page height. Use <span class='field-value'>-1</span> for a single page PDF. The safe maximum is <span class='field-value'>200in</span> otherwise some PDF viewers may be unable to open the PDF. The value must be -1 or specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setPageDimensions($width, $height) {
@@ -1233,12 +1233,12 @@ class HtmlToPdfClient {
     /**
     * Set the output page top margin.
     *
-    * @param top Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param top The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setMarginTop($top) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $top))
-            throw new Error(create_invalid_value_message($top, "setMarginTop", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_margin_top"), 470);
+            throw new Error(create_invalid_value_message($top, "setMarginTop", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_margin_top"), 470);
         
         $this->fields['margin_top'] = $top;
         return $this;
@@ -1247,12 +1247,12 @@ class HtmlToPdfClient {
     /**
     * Set the output page right margin.
     *
-    * @param right Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param right The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setMarginRight($right) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $right))
-            throw new Error(create_invalid_value_message($right, "setMarginRight", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_margin_right"), 470);
+            throw new Error(create_invalid_value_message($right, "setMarginRight", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_margin_right"), 470);
         
         $this->fields['margin_right'] = $right;
         return $this;
@@ -1261,12 +1261,12 @@ class HtmlToPdfClient {
     /**
     * Set the output page bottom margin.
     *
-    * @param bottom Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param bottom The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setMarginBottom($bottom) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $bottom))
-            throw new Error(create_invalid_value_message($bottom, "setMarginBottom", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_margin_bottom"), 470);
+            throw new Error(create_invalid_value_message($bottom, "setMarginBottom", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_margin_bottom"), 470);
         
         $this->fields['margin_bottom'] = $bottom;
         return $this;
@@ -1275,12 +1275,12 @@ class HtmlToPdfClient {
     /**
     * Set the output page left margin.
     *
-    * @param left Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param left The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setMarginLeft($left) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $left))
-            throw new Error(create_invalid_value_message($left, "setMarginLeft", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_margin_left"), 470);
+            throw new Error(create_invalid_value_message($left, "setMarginLeft", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_margin_left"), 470);
         
         $this->fields['margin_left'] = $left;
         return $this;
@@ -1300,10 +1300,10 @@ class HtmlToPdfClient {
     /**
     * Set the output page margins.
     *
-    * @param top Set the output page top margin. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
-    * @param right Set the output page right margin. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
-    * @param bottom Set the output page bottom margin. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
-    * @param left Set the output page left margin. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param top Set the output page top margin. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
+    * @param right Set the output page right margin. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
+    * @param bottom Set the output page bottom margin. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
+    * @param left Set the output page left margin. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setPageMargins($top, $right, $bottom, $left) {
@@ -1342,12 +1342,12 @@ class HtmlToPdfClient {
     /**
     * Set the top left X coordinate of the content area. It is relative to the top left X coordinate of the print area.
     *
-    * @param x Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt). It may contain a negative value.
+    * @param x The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt". It may contain a negative value.
     * @return The converter object.
     */
     function setContentAreaX($x) {
         if (!preg_match("/(?i)^0$|^\-?[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $x))
-            throw new Error(create_invalid_value_message($x, "setContentAreaX", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt). It may contain a negative value.", "set_content_area_x"), 470);
+            throw new Error(create_invalid_value_message($x, "setContentAreaX", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\". It may contain a negative value.", "set_content_area_x"), 470);
         
         $this->fields['content_area_x'] = $x;
         return $this;
@@ -1356,12 +1356,12 @@ class HtmlToPdfClient {
     /**
     * Set the top left Y coordinate of the content area. It is relative to the top left Y coordinate of the print area.
     *
-    * @param y Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt). It may contain a negative value.
+    * @param y The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt". It may contain a negative value.
     * @return The converter object.
     */
     function setContentAreaY($y) {
         if (!preg_match("/(?i)^0$|^\-?[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $y))
-            throw new Error(create_invalid_value_message($y, "setContentAreaY", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt). It may contain a negative value.", "set_content_area_y"), 470);
+            throw new Error(create_invalid_value_message($y, "setContentAreaY", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\". It may contain a negative value.", "set_content_area_y"), 470);
         
         $this->fields['content_area_y'] = $y;
         return $this;
@@ -1370,12 +1370,12 @@ class HtmlToPdfClient {
     /**
     * Set the width of the content area. It should be at least 1 inch.
     *
-    * @param width Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param width The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setContentAreaWidth($width) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $width))
-            throw new Error(create_invalid_value_message($width, "setContentAreaWidth", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_content_area_width"), 470);
+            throw new Error(create_invalid_value_message($width, "setContentAreaWidth", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_content_area_width"), 470);
         
         $this->fields['content_area_width'] = $width;
         return $this;
@@ -1384,12 +1384,12 @@ class HtmlToPdfClient {
     /**
     * Set the height of the content area. It should be at least 1 inch.
     *
-    * @param height Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param height The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setContentAreaHeight($height) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $height))
-            throw new Error(create_invalid_value_message($height, "setContentAreaHeight", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_content_area_height"), 470);
+            throw new Error(create_invalid_value_message($height, "setContentAreaHeight", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_content_area_height"), 470);
         
         $this->fields['content_area_height'] = $height;
         return $this;
@@ -1398,10 +1398,10 @@ class HtmlToPdfClient {
     /**
     * Set the content area position and size. The content area enables to specify a web page area to be converted.
     *
-    * @param x Set the top left X coordinate of the content area. It is relative to the top left X coordinate of the print area. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt). It may contain a negative value.
-    * @param y Set the top left Y coordinate of the content area. It is relative to the top left Y coordinate of the print area. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt). It may contain a negative value.
-    * @param width Set the width of the content area. It should be at least 1 inch. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
-    * @param height Set the height of the content area. It should be at least 1 inch. Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param x Set the top left X coordinate of the content area. It is relative to the top left X coordinate of the print area. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt". It may contain a negative value.
+    * @param y Set the top left Y coordinate of the content area. It is relative to the top left Y coordinate of the print area. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt". It may contain a negative value.
+    * @param width Set the width of the content area. It should be at least 1 inch. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
+    * @param height Set the height of the content area. It should be at least 1 inch. The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setContentArea($x, $y, $width, $height) {
@@ -1457,12 +1457,12 @@ class HtmlToPdfClient {
     /**
     * Set the header height.
     *
-    * @param height Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param height The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setHeaderHeight($height) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $height))
-            throw new Error(create_invalid_value_message($height, "setHeaderHeight", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_header_height"), 470);
+            throw new Error(create_invalid_value_message($height, "setHeaderHeight", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_header_height"), 470);
         
         $this->fields['header_height'] = $height;
         return $this;
@@ -1510,12 +1510,12 @@ class HtmlToPdfClient {
     /**
     * Set the footer height.
     *
-    * @param height Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).
+    * @param height The value must be specified in inches "in", millimeters "mm", centimeters "cm", or points "pt".
     * @return The converter object.
     */
     function setFooterHeight($height) {
         if (!preg_match("/(?i)^0$|^[0-9]*\.?[0-9]+(pt|px|mm|cm|in)$/", $height))
-            throw new Error(create_invalid_value_message($height, "setFooterHeight", "html-to-pdf", "Can be specified in inches (in), millimeters (mm), centimeters (cm), or points (pt).", "set_footer_height"), 470);
+            throw new Error(create_invalid_value_message($height, "setFooterHeight", "html-to-pdf", "The value must be specified in inches \"in\", millimeters \"mm\", centimeters \"cm\", or points \"pt\".", "set_footer_height"), 470);
         
         $this->fields['footer_height'] = $height;
         return $this;